shipment operations now work regarding to products

+++ src/main/java/pl/spomex/bazodan/shipment/ShipmentService.java
@@ -2,6 +2,8 @@ package pl.spomex.bazodan.shipment;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import pl.spomex.bazodan.product.Product;
+import pl.spomex.bazodan.product.ProductRepository;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -11,6 +13,9 @@ public class ShipmentService {
 
     @Autowired
     private ShipmentRepository shipmentRepository;
+
+    @Autowired
+    private ProductRepository productRepository;
 
     public List<Shipment> getAllShipments() {
         List<Shipment> shipments = new ArrayList<>();
@@ -23,14 +28,22 @@ public class ShipmentService {
     }
 
     public void addShipment(Shipment shipment) {
-        shipmentRepository.save(shipment);
+        updateShipment(shipment);
     }
 
     public void updateShipment(Shipment shipment) {
-        shipmentRepository.save(shipment);
+        Shipment newShipment = shipmentRepository.save(shipment);
+        for (Product product : newShipment.getProducts()) {
+            product.setShipment(newShipment);
+            productRepository.save(product);
+        }
     }
 
     public void deleteShipment(Integer id) {
-        shipmentRepository.deleteById(id);
+        Shipment shipment = shipmentRepository.findById(id).orElse(null);
+        if (shipment != null) {
+            shipment.getProducts().forEach(t -> productRepository.delete(t));
+            shipmentRepository.deleteById(id);
+        }
     }
 }
